+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 ## Unreleased
 
 🚨 *Breaking Changes*
-*  `sdk.WorkflowRun.get_logs()` now accepts TaskInvocationID instead of TaskRunID
+* `sdk.WorkflowRun.get_logs()` now accepts TaskInvocationID instead of TaskRunID
+* `sdk.WorkflowRun.get_artifacts()` doesn't accept any arguments anymore. Now, it returns all the artifacts produced by the tasks in the workflow.
+
 
 🔥 *Features*
+
 * `list_workflow_runs` added to the Public API. This lets you list the workflows for a given config, for example `sdk.list_workflow_runs("ray")` or `sdk.list_workflow_runs("prod-d")`.
 
 
@@ -19,7 +22,10 @@
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry task logs`.
 
 🐛 *Bug Fixes*
+
 * Fixed broken link on docs landing page.
+* Internal logs from Ray are no longer displayed.
+* Fixed the docstrings for `sdk.WorkflowRun.get_artifacts()`. It returns a dictionary with `TaskInvocationID` as keys and whatever the task returns as values.
 
 
 *Internal*